requirements.txt
Django==3.0.7
djangorestframework==3.11.0
django_compressor==2.4
django-countries==6.1.2
django-crispy-forms==1.9.1
django-environ==0.4.5
django_storages[azure]==1.9.1
gunicorn==20.0.4
markdown==3.2.2
psycopg2-binary==2.8.5
